Complete rework into ProcessorSink interface:
This new interface is a mix of (Flux|Mono)Sink and common reactor
processor APIs extracted from FluxProcessor and MonoProcessor.

It has additional specializations for Mono and Flux, which expose the
relevant sink methods and `asXXX` conversion methods.

The implementations simply delegate to the various concrete
FluxProcessors or MonoProcessor and create a FluxSink or MonoSink on
behalf of the user to delegate to for the sink APIs.